Detect local template files
If a file `FILE` is to be generated, `drom` will check if a file
`FILE.drom` exists. If it exists, its content will be used as a
template to generate `FILE`, overwriting the content that was found
in the skeleton tree. Such files can be created from template files
from drom-share, but also directly from `_drom/skeleton/` in your
project, where a copy of the skeleton is generated during
`drom project`.